+++ src/print.rs
@@ -3,37 +3,26 @@
 use {
     crate::{
         app::*,
-        display::{DisplayableTree, Screen},
+        display::Screen,
         errors::ProgramError,
         launchable::Launchable,
-        skin::{ExtColorMap, PanelSkin, StyleMap},
+        skin::{PanelSkin, StyleMap},
         tree::Tree,
     },
     crossterm::tty::IsTty,
     pathdiff,
     std::{
-        fs::OpenOptions,
-        io::{self, Write, stdout},
+        io::{self, stdout},
         path::Path,
     },
 };
 
-fn print_string(string: String, con: &AppContext) -> io::Result<CmdResult> {
+fn print_string(string: String, _con: &AppContext) -> io::Result<CmdResult> {
     Ok(
-        if let Some(ref output_path) = con.launch_args.file_export_path {
-            // an output path was provided, we write to it
-            let f = OpenOptions::new()
-                .create(true)
-                .append(true)
-                .open(output_path)?;
-            writeln!(&f, "{}", string)?;
-            CmdResult::Quit
-        } else {
-            // no output path provided. We write on stdout, but we must
-            // do it after app closing to have the desired stdout (it may
-            // be the normal terminal or a file, or other output)
-            CmdResult::from(Launchable::printer(string))
-        }
+        // We write on stdout, but we must do it after app closing
+        // to have the desired stdout (it may be the normal terminal
+        // or a file, or other output)
+        CmdResult::from(Launchable::printer(string))
     )
 }
 
@@ -54,7 +43,7 @@ pub fn print_paths(sel_info: &SelInfo, con: &AppContext) -> io::Result<CmdResult
 }
 
 fn relativize_path(path: &Path, con: &AppContext) -> io::Result<String> {
-    let relative_path = match pathdiff::diff_paths(path, &con.launch_args.root) {
+    let relative_path = match pathdiff::diff_paths(path, &con.initial_root) {
         None => {
             return Err(io::Error::new(
                 io::ErrorKind::Other,
@@ -88,53 +77,23 @@ pub fn print_relative_paths(sel_info: &SelInfo, con: &AppContext) -> io::Result<
     print_string(string, con)
 }
 
-fn print_tree_to_file(
-    tree: &Tree,
-    screen: Screen,
-    file_path: &str,
-    ext_colors: &ExtColorMap,
-) -> Result<CmdResult, ProgramError> {
-    let no_style_skin = StyleMap::no_term();
-    let dp = DisplayableTree::out_of_app(
-        tree,
-        &no_style_skin,
-        ext_colors,
-        screen.width,
-        (tree.lines.len() as u16).min(screen.height),
-    );
-    let mut f = OpenOptions::new()
-        .create(true)
-        .append(true)
-        .open(file_path)?;
-    dp.write_on(&mut f)?;
-    Ok(CmdResult::Quit)
-}
-
 pub fn print_tree(
     tree: &Tree,
     screen: Screen,
     panel_skin: &PanelSkin,
     con: &AppContext,
 ) -> Result<CmdResult, ProgramError> {
-    if let Some(ref output_path) = con.launch_args.file_export_path {
-        // an output path was provided, we write to it
-        print_tree_to_file(tree, screen, output_path, &con.ext_colors)
+    // We write on stdout, but we must do it after app closing to have the normal terminal
+    let show_color = con.launch_args.color.unwrap_or_else(|| stdout().is_tty());
+    let styles = if show_color {
+        panel_skin.styles.clone()
     } else {
-        // no output path provided. We write on stdout, but we must
-        // do it after app closing to have the normal terminal
-        let show_color = con.launch_args
-            .color
-            .unwrap_or_else(|| stdout().is_tty());
-        let styles = if show_color {
-            panel_skin.styles.clone()
-        } else {
-            StyleMap::no_term()
-        };
-        Ok(CmdResult::from(Launchable::tree_printer(
-            tree,
-            screen,
-            styles,
-            con.ext_colors.clone(),
-        )))
-    }
+        StyleMap::no_term()
+    };
+    Ok(CmdResult::from(Launchable::tree_printer(
+        tree,
+        screen,
+        styles,
+        con.ext_colors.clone(),
+    )))
 }